6.1.17 Выносит общую логику интерактивного компонента в отдельный класс Smart.
